% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStanDataForFitting.R
\name{createStanDataForFitting}
\alias{createStanDataForFitting}
\title{Create list of data for Stan fitting}
\usage{
createStanDataForFitting(
  win,
  loss = NULL,
  choice,
  numDecks,
  utility,
  updating,
  temperature,
  pars = NULL,
  reg = 1,
  scale = 0.01
)
}
\arguments{
\item{win}{Vector or wins}

\item{loss}{Vector of losses}

\item{choice}{Integer vector of deck choices}

\item{numDecks}{Integer. Number of decks}

\item{utility}{Name of utility function}

\item{updating}{Name of updating function}

\item{temperature}{Name of temperature function}

\item{pars}{A named list of parameters. See details.}

\item{reg}{Positive regularization parameter. A Beta(reg,reg) prior is placed
over the range of each parameter, with larger values of reg giving a tighter
concentration around the midpoint of the interval. Default value of 1 gives
a uniform prior.}

\item{scale}{Scaling factor for observed outcomes}
}
\value{
A list of data for stan
}
\description{
Create list of data for Stan fitting
}
\details{
Argument "pars" is a named list with fields:
\itemize{
 \item{"utility": }{A named list of parameter values. Names must correspond to
 the parameters of the utility function specified by argument "utility".
 Each parameter must be a vector of length 2, giving lower and upper bounds.}
 \item{"updating": }{See above.}
 \item{"temperature": }{See above.}
}
}
