% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStanDataForSimulation.R
\name{createStanDataForSimulation}
\alias{createStanDataForSimulation}
\title{Create list of data for Stan simulation}
\usage{
createStanDataForSimulation(deck, utility, updating, temperature, pars, scale)
}
\arguments{
\item{utility}{Name of utility function}

\item{updating}{Name of updating function}

\item{temperature}{Name of temperature function}

\item{pars}{A named list of parameters. See details.}

\item{decks}{Names list with entries "win" and "loss"}
}
\value{
A list of data for stan
}
\description{
Create list of data for Stan simulation
}
\details{
Argument "pars" is a named list with fields:
\itemize{
 \item{"utility": }{A named list of parameter values. Names must correspond to
 the parameters of the utility function specified by argument "utility".}
 \item{"updating": }{See above.}
 \item{"temperature": }{See above.}
}
}
