% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitIGT.R
\name{fitIGT}
\alias{fitIGT}
\title{Fit model to IGT data}
\usage{
fitIGT(
  choice,
  win,
  loss = NULL,
  numDecks,
  utility,
  updating,
  temperature,
  pars,
  scale = 0.01,
  reg = 1,
  method = "optimizing",
  returnStanfit = F,
  ...
)
}
\arguments{
\item{choice}{Integer vector of deck choices}

\item{win}{Vector of wins. If only net outcomes are observed, they should be provided here.}

\item{loss}{Vector of losses. May be omitted if only net outcomes are observed.}

\item{numDecks}{Integer. Number of decks}

\item{utility}{Name of utility function}

\item{updating}{Name of updating function}

\item{temperature}{Name of temperature function}

\item{pars}{A named list of parameters. See details.}

\item{scale}{Scaling factor for observed outcomes. Typically, outcomes are divided
by 100, corresponding to scale = .01}

\item{reg}{Positive regularization parameter. A Beta(reg,reg) prior is placed
over the range of each parameter, with larger values of reg giving a tighter
concentration around the midpoint of the interval. Default value of 1 gives
a uniform prior.}

\item{method}{Either "optimization" for MAP (MLE when reg = 1), or "sampling"
for full posterior sampling via Stan's No-U-Turn-Sampler (NUTS).}

\item{returnStanfit}{Whether to return the full stanfit object.}

\item{...}{Optional arguments provided to rstan::sampling()}
}
\value{
Nothing. Generates an error message in the event of incorrect input.
}
\description{
Fit model to IGT data either by MLE/MAP, or full posterior sampling
}
\details{
Argument "pars" is a named list with fields:
\itemize{
 \item{"utility": }{A named list of parameter values. Names must correspond to
 the parameters of the utility function specified by argument "utility".
 Each parameter must be a vector of length 2, giving lower and upper bounds.}
 \item{"updating": }{See above.}
 \item{"temperature": }{See above.}
}
Note that any or all parameters may be omitted, in which case default bounds
will be selected based on conventions in the literature. If any of "utility",
"updating", or "temperature" are missing, default bounds will be selected for
all parameters.

The user may pass optional arguments to rstan::sampling(). In the absence of
such arguments, the function uses the defaults, which at the time of writing
specify four chains of 2000 samples for full posterior sampling.
}
