% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatLikelihoodOutput.R
\name{formatLikelihoodOutput}
\alias{formatLikelihoodOutput}
\title{Organize likelihood generated quantities}
\usage{
formatLikelihoodOutput(draws, paramValues)
}
\arguments{
\item{draws}{A draws_array object returned by a simulate_igt model}

\item{paramValues}{A draws_array object returned by createStanParamsForSimulation()}
}
\value{
A named list containing the log-likelihood for the provided model parameters
}
\description{
Organize likelihood generated quantities
}
