% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowas predictions.R
\name{GetPrediction}
\alias{GetPrediction}
\title{Make prediction on the sales price}
\usage{
GetPrediction()
}
\value{
The prediction result for the column will be returned
}
\description{
The function will bind the input data from the user to the training data
It then uses the training model to predict the sales price for that row
}
\examples{
GetPrediction()
}
