% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choiceFuns.R
\name{igtChoice}
\alias{igtChoice}
\title{Choice functions}
\usage{
igtChoice(V, par, trial = NULL, FUN = NULL)
}
\arguments{
\item{V}{Vector of deck valuations.}

\item{par}{A named list of parameters. See details.}

\item{trial}{Current trial number. Only needed for FUN = vartemp.}

\item{FUN}{The choice function to be used. See details.}
}
\value{
A vector of choice probabilities.
}
\description{
Implements several learning functions commonly used in models of the IGT
}
\details{
All choice functions require a single parameter c.
Allowable choice functions are
\itemize{
 \item{"fixedtemp": }{Temperature is equal to 3^c - 1.}
 \item{"vartemp": }{Temperature is equal to (trial/10)^c.}
}
}
