% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learningFuns.R
\name{igtLearning}
\alias{igtLearning}
\title{Learning functions}
\usage{
igtLearning(u, V, choice, par, FUN = NULL)
}
\arguments{
\item{u}{A scalar. The observed utility.}

\item{V}{A vector of current deck valuations.}

\item{choice}{An integer indicating the current deck choice.}

\item{par}{A named list of parameters. See details.}

\item{FUN}{The learning function to be used. See details.}
}
\description{
Implements several learning functions commonly used in models of the IGT
}
\details{
All learning functions require a single learning rate parameter r.
Allowable learning functions are
\itemize{
 \item{"PV": }{The learning function used in the PVL model.}
 \item{"EV": }{The learning function used in the EV model.}
 \item{"Delta": }{The delta learning rule.}
}
}
