% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityfuns.R
\name{igtUtility}
\alias{igtUtility}
\title{Utility functions}
\usage{
igtUtility(x = NULL, g = NULL, l = NULL, par, FUN = NULL)
}
\arguments{
\item{x, g, l}{For utility functions which depend only on the net outcome, x is a scale/vector
of net outcomes. For utilities which depend on both the observed gain and loss (e.g. the EV utility),
g and l are scalars/vectors indicating the gain and loss.}

\item{par}{A named list of parameters. See details.}

\item{FUN}{The utility function to be used. See details.}
}
\description{
Implements several utility functions commonly used in models of the IGT
}
\details{
Allowable utility functions are
\itemize{
 \item{"PV": }{The prospect utility function. Depends only on the net outcome x,
 and requires a shape parameter A and a loss aversion parameter l}
 \item{"EV": }{The expectancy valence utility function. Depends on both the observed gain and loss (g,l),
 and requires a loss aversion parameter w.}
}
}
