% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iowa-package.R
\docType{package}
\name{iowa-package}
\alias{iowa-package}
\alias{iowa}
\title{The 'iowa' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
\seealso{
Useful links:
\itemize{
  \item \url{https://areshenk-rpackages.github.io/iowa/}
}

}
