% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodGT.R
\name{likelihoodIGT}
\alias{likelihoodIGT}
\title{Compute likelihood of an IGT model}
\usage{
likelihoodIGT(
  wins,
  losses,
  choices,
  numDecks,
  utility,
  updating,
  temperature,
  pars,
  scale = 0.01
)
}
\arguments{
\item{wins}{Vector of wins}

\item{losses}{Vector of losses}

\item{numDecks}{Number of decks}

\item{updating}{Name of updating function}

\item{temperature}{Name of temperature function}

\item{pars}{A named list of parameters. See details.}

\item{scale}{How to scale outcomes before passing to the model. Typically, outcomes
are divided by 100 (i.e. scale = .01).}

\item{choice}{Vector of deck choices}
}
\value{
A named list containing simulated IGT data.
}
\description{
Likelihood calculation for a specified model.
}
\details{
Argument "pars" is a named list with fields:
\itemize{
 \item{"utility": }{A named list of parameter values. Names must correspond to
 the parameters of the utility function specified by argument "utility".}
 \item{"updating": }{See above.}
 \item{"temperature": }{See above.}
}
}
