% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowas UI.R
\name{Logic}
\alias{Logic}
\title{Turning input data into a dataframe}
\usage{
Logic(input)
}
\arguments{
\item{input}{A selection of valus that matches a dataframe format should be checked}
}
\value{
The dataframe containing the inputs will be returned
}
\description{
The function will obatain the input from the front-end UI insert them into a dataframe
It tranforms the values of the dataframe to a useble format that matches the training data
}
