% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setStanParamsForSimulation.R
\name{setStanParamsForSimulation}
\alias{setStanParamsForSimulation}
\title{Create parameter object for Stan}
\usage{
setStanParamsForSimulation(utility, updating, temperature, pars)
}
\arguments{
\item{utility}{Name of utility function}

\item{updating}{Name of updating function}

\item{temperature}{Name of temperature function}

\item{pars}{A named list of parameters. See details.}
}
\value{
A list for Stan.
}
\description{
Internal function.
}
\details{
Argument "pars" is a named list with fields:
\itemize{
 \item{"utility": }{A named list of parameter values. Names must correspond to
 the parameters of the utility function specified by argument "utility".}
 \item{"updating": }{See above.}
 \item{"temperature": }{See above.}
}
}
