% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateIGT.R
\name{simulateIGT}
\alias{simulateIGT}
\title{Simulate performance of an IGT model}
\usage{
simulateIGT(
  deck = "bechara",
  n,
  utility,
  updating,
  temperature,
  pars,
  scale = 0.01,
  fullOutput = FALSE
)
}
\arguments{
\item{deck}{Either the name of an included deck or a named list
with matrices entries "win" and "loss".}

\item{n}{Number of subjects to simulate}

\item{updating}{Name of updating function}

\item{temperature}{Name of temperature function}

\item{pars}{A named list of parameters. See details.}

\item{scale}{How to scale outcomes before passing to the model. Typically, outcomes
are divided by 100 (i.e. scale = .01).}

\item{fullOutput}{Whether to output internal state variables (i.e. deck valuations)}
}
\value{
A named list containing simulated IGT data.
}
\description{
Simulate IGT performance using a specified model.
}
\details{
Argument "pars" is a named list with fields:
\itemize{
 \item{"utility": }{A named list of parameter values. Names must correspond to
 the parameters of the utility function specified by argument "utility".}
 \item{"updating": }{See above.}
 \item{"temperature": }{See above.}
}
}
\seealso{
[importDeck()] for available deck structures, and more information
about the structure of a deck object
}
