% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifyFittingBounds.R
\name{verifyFittingBounds}
\alias{verifyFittingBounds}
\title{Verify provided fitting bounds}
\usage{
verifyFittingBounds(utility, updating, temperature, pars = NULL)
}
\arguments{
\item{utility}{Name of utility function}

\item{updating}{Name of updating function}

\item{temperature}{Name of temperature function}

\item{pars}{A named list of parameters. See details.}
}
\value{
Generates an error message in the event of incorrect input. Otherwise,
returns a copy of pars, possibly with missing bounds set to their defaults.
}
\description{
Internal function. Verify that fitting parameters are correct, and correspond to implemented models.
}
\details{
Argument "pars" is a named list with fields:
\itemize{
 \item{"utility": }{A named list of parameter values. Names must correspond to
 the parameters of the utility function specified by argument "utility".}
 \item{"updating": }{See above.}
 \item{"temperature": }{See above.}
}
}
