% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{iowa-package}
\alias{iowa}
\alias{iowa-package}
\title{iowa: Modular reinforcement learning models for the Iowa gambling task}
\description{
cmdstanr models of the Iowa gambling task
}
\seealso{
Useful links:
\itemize{
  \item \url{https://areshenk-rpackages.github.io/iowa/}
}

}
\concept{help}
