% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateIGT.R
\name{simulateIGT}
\alias{simulateIGT}
\title{Simulate performance of IGT model}
\usage{
simulateIGT(
  deck,
  n,
  utility,
  updating,
  temperature,
  pars,
  scale = 0.01,
  fullOutput = FALSE
)
}
\arguments{
\item{utility}{Name of utility function}

\item{updating}{Name of updating function}

\item{temperature}{Name of temperature function}

\item{pars}{A named list of parameters. See details.}
}
\value{
Nothing. Generates an error message in the event of incorrect input.
}
\description{
Internal function. Verify that simulation parameters are correct, and correspond to implemented models.
}
\details{
Argument "pars" is a named list with fields:
\itemize{
 \item{"utility": }{A named list of parameter values. Names must correspond to
 the parameters of the utility function specified by argument "utility".}
 \item{"updating": }{See above.}
 \item{"temperature": }{See above.}
}
}
